Add SPQ tests to mdtest
A Markdown fenced code block with mdtest-spq as the first word of its
info string contains an SPQ test.  The content of the block must
comprise three sections, each preceeded by one or more "#"-prefixed
lines.  The first section contains an SPQ program, the second contains
input provided to the program when the test runs, and the third contains
the program's expected output.