lighter weight type context management
This commit makes a big performance improvement for type context
management by removing the need to serialize and store type values
for every type in the context.  This was especially problematic
for large (wide and deep) types where the system would reserialize every
node in the type hierarchy from bottom to top so there was a quadratic
effect going on.  We never noticed this for small types but for the
large, nested values of CSUP metadata the problem became painfully
obvious.

Part of the fix is to remove super.Mapper, which was based on mapping
types by serialized type value, and replace it with a direct mapper
in the BSUP reader where local type IDs are mapped directly to
the shared super.Context.  As part of this change, we cleaned up
the mapper cache by creating the super.TypeFetcher interface and
arraging for super.Context, zngio.Decoder, and super.TypeCache to
all implement this interface.  This way the bufferfilter can simply
take super.TypeFetcher interface value and not worry that the type
mappings are handled in the BSUP decoder.

Types are now interned by using type fingerprints based on local
type IDs instead of large, serialized type values.  There is a dedicated
map in the super.Context for each complex type category but the maps
are allocated only when needed.

These changes mean that all values in the query runtime must be typed
from the shared query super.Context.  This was generally the case before,
but we found a problem in network_of where error values were using
zson.MarshalZNG, which created a temporary super.Context and caused
mixed-context values to be created and thus foiled the type interning
in the shared context.  This has been fixed by passing the context
to the marshaler.

Finally, since we were touching the code, we added some circuit-breaker
checks on the number of union types and number of enum symbols in the
BSUP reader.