remove super.Mapper from flattener
This commit changes the use of super.Mapper in the flattener
to a simple hash table.  Mapper has unnecessary overhead for this
use case (for example, computing type values to translate between
contexts when this code used same context as the output) and a simple
map implementation suffices here.  Also, we plan to change the
super.Mapper API in a subsequent PR and its use here would otherwise
complicate that change.