lazy type decoding for CSUP
This commit addresses a scaling challenge for CSUP with respect
to large numbers of types by arranging for the per-type metadata
to be decoded on an as-needed basis.  To do so, we changed the
metadata representation from a single super value to an array of
flattened metadata records indexed by ID.  This allows us to
incrementally unmarshal and build the shadow vectors fow only the
types needed by a query.  Additionally, the metadata object filter
now performs projection too so that only the metadata values from
the types needed are deserialized and acted upon by the metadata
filter.

These changes affect the CSUP file format so we bumped its version
number from 8 to 9.

The code that unmarshals the shadow vectors is not currently reentrant.
If/when we want to allow concurrent updates (e.g., vcache used by
parallel lake requests) we will need to invoke the proper locking protocol.

Some rough perf measurements indicate a 5X speedup for the
Bluesky Million data set on a simple query projecting a single column.