rename vector.NewView to vector.Pick
This commit renames NewView and NewInverseView to Pick and ReversePick
since NewView only creates a View when it has to.  This better reflects
the verb of what's going on and makes the code easier to read.

We also created a real NewView function when the intention is to
create a vector.View.  These changes exposed a couple places in the
code where the intention was to create a View and not do a Pick so
these call sites were left unchanged.