rework vcache shadow subsystem
This commit refactors the vcache shadow design so that incremental
building and loading of shadow values works properly under concurrency.
The design has been improved by doing most the work through two key
methods on the shadow interface: unmarshal and project, both taking
a projection argument to prune the data loaded to just the projection
indicated.

This design scales well with large numbers of types because only the
types and CSUP metadata that is needed is ever decoded.

For each shadow type, there is a loader type that implements a set of
new interfaces in the vector runtime for incrementally loading vector
data based on usage, i.e., only data that is needed by a query is
actually loaded.  This logic is not yet hooked up to the runtime but
will be utilized in the forthcoming lazy-vectors PR.

We removed quite a bit of concurrency that was enabled by the previous
design.  This may help with processor cache locality for fast storage
access like local SSD but will definitely hurt high-latency access
like cloud-based object store.  We can pretty easily add back
concurrent reads when warranted.