add SQL VALUES clause and column aliases
This commit adds support for a VALUES clause as well as an optional
columns specification on table aliases.  The values listed in the
values clause must all be uniformly typed.  When polymorphic data
is desired, yield can be used inside of a SQL pipe instead of this
values clause.  VALUES is not ANSI SQL but provides compatibility with
PostgreSQL and duckdb.