proper double-quoted SQL identifiers
This commit cleans up the handling of SQL identifiers in double quotes.
In SQL expressions, double-quoted strings are always now interpreted
as identifiers.  In pipe expressions, double-quoted strings are string
literals.  In either context, single-quoted strings are always string
literals and backtick-quoted strings are always identifiers.

To do this, we added a new AST expression type to represent double-quoted
strings so that this version of a string literal can be differentiated
in the semantic pass and handled according to the rules above.

We also updated the grammar to allow backtick-quoted and double-quoted
strings to be on the RHS of a dot operator.