Split vector.Dynamic.TagMap into ForwardTagMap and ReverseTagMap
A call to vector.Dynamic.TagMap creates both the forward and reverse tag
maps, but profiling shows that for a given vector, often only one is
used.  To avoid the wasted work (especially allocations) of creating the
other, remove the vector.TagMap type and split the Dynamic.TagMap method
into separate ForwardTagMap and ReverseTagMap methods.