new docs infrastructure
This commit is the first step toward a new docs design based
on mdbook from the Rust eco-system.  It also implements a better
playground for query examples allowing inputs, editing of the query
text, and copy-to-clipboard of the equivalent shell command.

The SUMMARY.md file defines the outline for the docs that are under
development.  Subsequent PRs will post updates to this outline
to fill out the book.

Co-authored-by: Noah Treuhaft <noah.treuhaft@gmail.com>
Co-authored-by: James Kerr <jkerr838@gmail.com>