remove fat arrows
This commit removes fat arrows (=>) from SuperSQL and replaces
them with parenthesised query fragments.

As part of this, we also removed the from (=> ... => ...)
syntax where one can simply instead use fork with embedded froms.
The deprecated forms of from (get/pool/file) have also been
removed.

This change introduced a potential ambiguity for over so we added
the keyword "into" to replace "=>" for over operators with lateral
scopes.  The "into" concept will be reused when we add other
forms of lateral scopes including partition-by, range windows,
sliding windows, and so forth.

We did not carefully update the docs subdirectory as all of these
changes will be accurately and comprehensively documented in the
upcoming book branch.