change union syntax to use pipe(|) instead of comma separator
This commit changes the SuperSQL grammar and SUP format to use
the pipe symbol for unions instead of parenthesized comma-separated
list of types.  This also required a new variation of a embedded
named type inside of a union where you can parenthesize the type,
e.g., to disambigate foo=int64|string.  If foo=int64 should be inside
of the union then the syntax is foo=(int64)|string.  Also, unions can
be parenthesized when they are inside another union, e.g.,
int64|(uint8|int8)|string.

Comprehensive docs updates are coming in the book branch.