normalize text type and from/load args in SuperSQL
This commit generalizes the syntax for handling optional
arguments used by "from" and "load" into a generic syntax of
parenthesized named/value pairs.  This is consistent
with other SQLs for specifying the serialization format
of external entities.

We also clarified the rule for text parameters (e.g.,
the from target, argument values, etc) and changed
ast.Name to ast.Text, where the PEG rule for a SUP name
is now called Name and reuses the ast.Text node.

This should be easier to maintain and to reuse for
future operators that take optional arguments.