automatically promote a user-op in an expression to a subquery
This commit changes the handling of user-ops referenced in expressions
from an error to a subquery.  When the user-op begins with a from,
then the subquery is spliced into the expression as an uncorrelated query.
Otherwise, it is dependenent on its input and the user-op is expanded
into the pattern (values this | userOp(params) | collect(this)).