new syntax for user operator declarations and invocations
This commit improves the syntax for user operators to better align
with built-in operators and distinguish their syntax from user
functions.

The new syntax for op declarations is:

     op name arg, arg, ... : ( ... )

and invoking an operator is now:

    call name arg, arg, ...

with a short-cut where you can drop the "call" as a shortcut:

    name arg, arg, ...

(as long as the shortcut is not a single operator with no arguments
inside of a subquery)

Updated documentation is coming in a subsequent PR from the book branch.