add lambdas and function references
This commit adds lambdas and function references to SuperSQL.
Lambdas are not first class as they cannot be referenced in
generic functions but they are callable and they are passable
to user operators and the map function.

We updated the docs for the map() function to explain lambdas
and function references but more comprehensive updates are
forthcoming from the book branch.

The way the function code is structured now does not easily allow
for passing lambdas to functions so this is not currently supported.
This can change if we update the model to look more like the
operator calling conventions where we could resolve and expand
lambas at compile time within function bodies, but this will be
deferred for now.