compiler/semantic: try to resolve ast.DoubleQuote in schema
When a schema is available, analyzer.semDoubleQuote should try to
resolve ast.DoubleQuote.Text in the schema.  It attempts to do this by
creating an ast.ID and passing that to semID, but this is incorrect
because schema resolution for an ast.ID happens in semExpr, not semID.
Call semExpr instead.

This bug is masked by another bug in the optimizer.  Rather than add an
explicit test for this bug, we should arrange to run the existing tests
without the optimizer.