compiler/optimizer: fix bug in liftFilterOps
liftFilterOps can produce a DAG in which a dag.Filter and dag.Values
share a dag.Expr.  If the shared Expr is later modified on behalf of one
operator, the change will incorrectly affect the other.  Fix by coping
any Expr from the dag.Values that is substituted into the dag.Filter.

I haven't been able to craft a test for this but need to fix it in order
to fix a mergeValuesOps bug that transforms "values {a:this} | values b"
to "values b" instead of "values error('missing')".