change Fn* naming pattern to *Func* pattern
This commit changes the Fn-style naming pattern that was used for
AST data structures to capture the concept of a SuperSQL fn to
a more lexically consistent Func-style naming pattern to be consistent
with any entity in the go source code being a function.  While
this comingles the two different concepts, it means there is only one
style of naming function entities so it is easier to remember.

There are lots of places in the existing code base that use Function
and Fn instead of Func.  This should be updated in a separate PR.