parser: Add back left recursion for DerefExpr
Pull-request #6162 attempted to get rid of dependence on the
experiemental left-recursion feature of pigeon by not using
left recursion for dereference expressions. The change had the
unexpected side-effect of adversely affecting the parsing times
of some queries. Rollback those changes and go back to using left
recursion for dereference expressions.