add syntactic sugar for array-wrapped subquery
This commit adds logic to the semantic pass to convert a subquery
that is immediately embedded in an array expression to that same
subquery followed by collect(this).  This makes it easier to write
queries that expect multiple values as output.  This will also help
with a subsequent PR that will add support for calling user ops
by name in an expression without having to put them in subquery parens.

Documentation for this syntax will be forthcoming from the book branch.