fix "super compile -dag/-O" panic on from op with Parquet file
When the semantic analyzer encounters a from operator with a Parquet
file as its argument, it tries to build a list of top-level fields in
that file.  For "super compile", this causes a panic because the file is
accessed through compiler/semantic.analyzer.env, which is nil.  Fix that.

Note that if a Parquet file is missing or otherwise inaccessible, its
schema won't be available.