compiler/semantic: do not share dag.Expr for ast.Between
semExpr creates a DAG for ast.Between that contains two references (in
two different nodes) to a single dag.Expr.  This causes a bug in
optimizer.liftFilterOps because a change through either reference is
unexpectedly visible through the other.  Fix this in semExpr by creating
two copies of the dag.Expr.