allow functions/lambdas to be passed as args to functions
This commit adds support for function arguments to functions.
Since functions are not first-class, the approach here is to do
a compile-time analysis of passed-in functions and create variations
of each function that takes different passed-in functions.  The
variation still calls the reference function by name by does so
in a static fashion, which means each variation of static binding
must be unraveled to a separate function definition in the DAG.
This is done in two passes in the semantic phase, where the first
pass decorages the DAG with function references and the second pass
creates the variations based on the actual references passed in.

We also changed the syntax so that user functions are referenced
as values simply with their name while built-in functions use
the &id syntax, e.g., myFunction vs. &upper.

We also fixed a bug where const decls could override the name of
a parameter when the parameter should override the const.  There's
a new test to cover this.

These changes also meant that ast.MapCall could be removed since
ast.Call now covers this use case.  We left dag.MapCall in place for
now but if/when we add more built-ins that take function args,
we can generalize dag.MapCall and extend the runtime package to
support different variations of such functions.