support for restartable dag.FileScans so repeated from-subqueries work
This commit adds support for restarting file scans so that repeated
from-subqueries involving file paths work. This is a temporary
solution to make progress on docs and testing.  It does not work
for databases and fails silently for non-seekable streams.

A more comprhensive solution will come in a future PR.