follow done protocol in hashjoin
I'm not 100% on this but it appears the hashjoin operator is not
following the done protocol precisely and calling its parent at
least an extra time after its already done.  In this new code, this
causes a restart so the upstream op is live again but the hashjoin
is still draining, which causes a path through the code to reference
the nil hits table.  This commit fixes the problem by wrapping the
bufPuller with another done bool and a strict adherence to the done
protocol so it can be called multiple times after the parent is done
without restarting the parent.  This is ok because the top hashjoin
object will reallocate the entire entity on restart and with this
new logic, the join parents should experience correct behavior and
restart and the realloc.