unify SQL and pipe name inference and simplify single by-only agg output
This commit unifies the way l-values are inferred for expressions
that omit the LHS of a field assignment or the AS clause in a select
SELECT expression.  ANSI SQL does not have specific guidance on how
to infer arbitrary expressions so we blended in some of the existing
logic from the pipe syntax.

We also changed how "this" is handled.  Instead of raising an error,
the l-value becomes "that".  We also adapted the heuristic for a single
column agg with no groupings to work also for a single grouping with
no aggs (i.e., emitting the values instead of a single-column record).
These two changes  means that "by this" now works and behaves like distinct.

These changes also means that there's more typing with operators like
cut when you want to preserve the object path, e.g., cut x.y.z:=x.y.z,
but less typing when you don't, e.g., cut x.y.z to mean cut z:=x.y.z.

Overall, this improves the consistency and orthogonality of the language.