back off on cut changes
Per our discussion, I tried to change cut to take path expressions
instead of assignments but this had an impact on the optimizer tests.
When changing cut to values in these tests, they were no longer
optimized the same way, which means the optimizer needs some work.
It needed this work anyway, but rather than do this here, I backed out
some of the cut changes and will defer the updates to a subsequent PR.