runtime/vam: return error on cast of invalid UTF-8 bytes to string
The sequential runtime returns a structured error when a bytes value
containing invalid UTF-8 is cast to a string.  Do the same in the
vector runtime.

The change in runtime/vam/expr/cast.castToString implements the new
behavior.  The remaining changes just add a msgSuffix parameter to
errCastFailed so castToString can produce the same error message as the
sequential runtime.

There are some possible optimizations here but they can wait until
castToString appears in CPU profiles.