add error check for multi-column SQL subquery results
This commit adds a dynamic error check for SQL subqueries to produce
an error when the subquery result has multiple columns.  It passes
through error values unmodified so that subquery errors are not mysteriously
converted to error("missing").

We also fixed a bug in the semantic analyzer to unravel subquery schemas
for select * and added test coverage for this.