integrate resolver's lambda unraveling into semantic translator pass
This commit integrates the lambda unraveling technique used in the
resolver into the first-pass of the semantic translator.  This is
possible by indexing the lambda permutation using decl IDs instead of
call tags based on the intuition that the only thing that can give
rise to variation in unraveled combinations is the declared function,
i.e., at the outermost call site, each variation of lamba-parameterized
function depends only on the function passed in, which are uniquely
defined by their ID not their unraveled instances.

This means the evaluator can simply traverse an expression to
determine if it's constant and then dagen it.  There is no need
anymore to re-enter the resolver.

This is also the key step needed to integrate type checking into
the translator, which will then allow us to utilize dataflow-analyzed
fused types from pipe queries in schema-column resolution for SQL
queries that include pipe subqueries.  This will be forthcoming in a
subsequent PR.