fix array subqueries in SQL expressions
This commit fixes a problem with the parser with respect to array
subqueries in SQL expressions as well as the semantic logic where
there is no need to extract column 1 from a presumed record in an array
subquery.  This logic will be further improved when the type checker
is integrated into the translator pass.