support column-id resolution in SQL pipe subqueries
This commit adds a pipe schema concept to the semantic pass so
that column references can be properly resolved for pipe subqueries
embedded as SQL table expressions.  It calls the type checker to
compute the super.Type of the pipe subquery and uses that type
to infer the available columns for SQL expressions.

This logic will need to be extended when we add support for correlated
subqueries, likely requiring type tracking to be more tightly integrated
with the translator pass (e.g., tracking the type all the way from the
input to the query instead of locally analyzing the pipe subquery).