Propogate SQL static schemas on join queries
This commit fixes and issue with the semantic pass for SQL queries
where join queries with static schemas would end up getting reported as
dynamic schemas. This would cause certain queries to lose out on potential
optimizations as well as inhibit future functionality that relies on
knowing if child schemas are static or not.