change GET /query/describe to send aggregate input keys
The GET /query/describe route sends a response containing the output
keys for aggregation operations but Zui is interested in the input keys,
so send those instead.

This wasn't a problem for Zui for most queries until #6263, which
changed the output key for "aggregate by a.b" from a.b (the path of the
input key) to b (the last path component of the input key).