big docs update primarily SQL
This commit adds SQL to the mdbook docs and makes a few other changes.

The pipe form of "from" has been unified with SQL from and updated
with a new style that better works with the left sidebar.  Here the
main topic of an md file is a single # at the top and each section
begins with ##.  We should do the rest of the docs in this style.
The SQL docs follow this style.

We also removed some trailing whitespace in a few places.